Refactor: move util functions to presentation layer and improve time formatting
- Moved utility extension functions from the `util` package to a new `presentation.util` package to better align with the app's architecture.
- The `util/Utils.kt` file has been deleted.
- Utility functions are now organized into separate files:
    - `DimensionExtensions.kt`: Contains `dpToPx`.
    - `ViewExtensions.kt`: Contains `show` and `setupWindowInsets`.
    - `TimeFormatExtensions.kt`: Contains `toMmSs`.
- All activities (`MainActivity`, `SearchActivity`, `SettingsActivity`, `PlayerActivity`, `MediaLibraryActivity`) and `TracksViewHolder` have been updated to use the new import paths for these extension functions.
- The `setupWindowInsets` function is now an extension function on `View`.
- Improved the `toMmSs` extension function:
    - It now uses a shared `SimpleDateFormat` instance with `Locale.ROOT` for consistent formatting.
    - An overload for `Long` has been added.
- Removed the local `toMmSs` implementation from `TrackDetailedInfoMapper` to use the new shared extension.
