+++ app/src/main/java/com/dimasla4ee/playlistmaker/presentation/util/DimensionExtensions.kt
@@ -1,0 +1,16 @@
+package com.dimasla4ee.playlistmaker.presentation.util
+
+import android.content.Context
+import android.util.TypedValue
+
+/**
+ * Converts a value in density-independent pixels (dp) to pixels (px).
+ *
+ * @param context The context to access resources and display metrics.
+ * @return The converted value in pixels.
+ */
+fun Float.dpToPx(context: Context): Float = TypedValue.applyDimension(
+    TypedValue.COMPLEX_UNIT_DIP,
+    this,
+    context.resources.displayMetrics
+)
